+++ javax/src/main/java/me/panpf/javax/io/FileTreeWalk.java
@@ -17,7 +17,7 @@
 package me.panpf.javax.io;
 
 import me.panpf.javax.collections.AbstractIterator;
-import me.panpf.javax.collections.Sequence;
+import me.panpf.javax.sequences.Sequence;
 import me.panpf.javax.util.*;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;