+++ javax/src/main/java/me/panpf/javax/io/Streamx.java
@@ -16,8 +16,8 @@
 
 package me.panpf.javax.io;
 
-import me.panpf.javax.collections.Sequence;
-import me.panpf.javax.collections.Sequencex;
+import me.panpf.javax.sequences.Sequence;
+import me.panpf.javax.sequences.Sequencex;
 import me.panpf.javax.lang.Charx;
 import me.panpf.javax.lang.Stringx;
 import me.panpf.javax.util.Action;