+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * :sparkles: Stringx adds asSequence method
 * :fire: MapBuilder rename build\* method
 * :fire: Rangex.until method now return \*Range
+* :sparkles: Add Sequencex